# docs/api/openapi.yaml
openapi: 3.1.0
info:
  description: クラウドファンディング型教育・実践開発システムのREST API（§6 API基本設計）
  title: CF-Training API
  version: v1
servers:
- url: /
paths:
...
  /api/v1/files/{fileId}/complete:
    post:
      operationId: complete
      parameters:
      - in: path
        name: fileId
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteUploadRequest"
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiEnvelopeCompleteUploadResponse"
          description: OK
      tags:
      - file-controller
...
  /api/v1/operations/payments/{paymentId}/reconcile:
    post:
      operationId: reconcile
      parameters:
      - in: path
        name: paymentId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiEnvelopeReconcileResponse"
          description: OK
      tags:
      - operations-controller
  /api/v1/operations/refunds:
    get:
      operationId: searchRefunds
      parameters:
      - in: query
        name: status
        required: false
        schema:
          type: string
      - in: query
        name: page
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - in: query
        name: size
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiEnvelopePageResultRefundListItem"
          description: OK
      tags:
      - operations-controller
  /api/v1/operations/refunds/{refundId}/retry:
    post:
      operationId: retry
      parameters:
      - in: path
        name: refundId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiEnvelopeRefundResponse"
          description: OK
      tags:
      - operations-controller
  /api/v1/operations/supports:
    get:
      operationId: searchSupports
      parameters:
      - in: query
        name: status
        required: false
        schema:
          type: string
      - in: query
        name: projectId
        required: false
        schema:
          type: string
      - in: query
        name: page
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - in: query
        name: size
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiEnvelopePageResultOperationsSupportListItem"
          description: OK
      tags:
      - operations-controller
  /api/v1/operations/supports/{supportId}/refunds:
    post:
      operationId: requestRefund
      parameters:
      - in: path
        name: supportId
        required: true
        schema:
          type: string
      - in: header
        name: Idempotency-Key
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRefundRequest"
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiEnvelopeRefundResponse"
          description: OK
      tags:
      - operations-controller
...
  /api/v1/projects/{projectId}/supports:
    post:
      operationId: request
      parameters:
      - in: path
        name: projectId
        required: true
        schema:
          type: string
      - in: header
        name: Idempotency-Key
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestSupportBody"
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiEnvelopeRequestSupportResponse"
          description: OK
      tags:
      - support-controller
...
components:
  schemas:
...
    ApiEnvelopeCompleteUploadResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/CompleteUploadResponse"
        meta:
          $ref: "#/components/schemas/ApiMeta"
...
    ApiEnvelopePageResultOperationsSupportListItem:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/PageResultOperationsSupportListItem"
        meta:
          $ref: "#/components/schemas/ApiMeta"
...
    ApiEnvelopePageResultRefundListItem:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/PageResultRefundListItem"
        meta:
          $ref: "#/components/schemas/ApiMeta"
...
    ApiEnvelopeReconcileResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/ReconcileResponse"
        meta:
          $ref: "#/components/schemas/ApiMeta"
    ApiEnvelopeRefundResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/RefundResponse"
        meta:
          $ref: "#/components/schemas/ApiMeta"
    ApiEnvelopeRequestSupportResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/RequestSupportResponse"
        meta:
          $ref: "#/components/schemas/ApiMeta"
...
    ApiMeta:
      type: object
      properties:
        correlationId:
          type: string
        timestamp:
          type: string
          format: date-time
...
    CompleteUploadRequest:
      type: object
      properties:
        sha256:
          type: string
          minLength: 1
          pattern: "[0-9a-fA-F]{64}"
      required:
      - sha256
    CompleteUploadResponse:
      type: object
      properties:
        downloadReference:
          type: string
        fileId:
          type: string
        status:
          type: string
...
    CreateRefundRequest:
      type: object
      properties:
        amount:
          type:
          - integer
          - "null"
          format: int64
          minimum: 1
        comment:
          type:
          - string
          - "null"
          maxLength: 2000
          minLength: 0
        reasonCode:
          type: string
          minLength: 1
      required:
      - reasonCode
...
    OperationsSupportListItem:
      type: object
      properties:
        amount:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
        paymentId:
          type:
          - string
          - "null"
        paymentStatus:
          type:
          - string
          - "null"
        projectId:
          type: string
        projectTitle:
          type: string
        status:
          type: string
          enum:
          - PENDING
          - PAID
          - PAYMENT_FAILED
          - CANCEL_REQUESTED
          - CANCELLED
          - REFUND_REQUESTED
          - REFUNDING
          - REFUNDED
          - REFUND_FAILED
        supportId:
          type: string
        supporterUserId:
          type: string
...
    PageResultOperationsSupportListItem:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/OperationsSupportListItem"
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
...
    PageResultRefundListItem:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/RefundListItem"
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
...
    ReconcileResponse:
      type: object
      properties:
        paymentId:
          type: string
        status:
          type: string
    RefundListItem:
      type: object
      properties:
        amount:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
        paymentId:
          type: string
        reasonCode:
          type: string
        refundId:
          type: string
        retryCount:
          type: integer
          format: int32
        status:
          type: string
          enum:
          - REQUESTED
          - PROCESSING
          - SUCCEEDED
          - RETRY_WAIT
          - FAILED
        supportId:
          type: string
        updatedAt:
          type: string
          format: date-time
    RefundResponse:
      type: object
      properties:
        refundId:
          type: string
        status:
          type: string
...
    RequestSupportBody:
      type: object
      properties:
        additionalAmount:
          type: integer
          format: int64
          minimum: 0
        contactEmail:
          type: string
          format: email
          minLength: 1
        quantity:
          type: integer
          format: int32
          maximum: 99
          minimum: 1
        rewardPlanId:
          type:
          - string
          - "null"
        termsAccepted:
          type: boolean
      required:
      - contactEmail
    RequestSupportResponse:
      type: object
      properties:
        paymentStatus:
          type: string
        statusUrl:
          type: string
        supportId:
          type: string
...